refactor: move CompletionRing to components/ui/
CompletionRing is a generic animated SVG progress ring — not
dashboard-specific. It belongs next to the other UI primitives so
it can be reused anywhere a percentage focal point is useful.

Single import update in user-profile.tsx. No behavior change.
