+++ components/dashboard/user-profile.tsx
@@ -2,7 +2,7 @@ import Link from "next/link"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import { AnimatedNumber } from "@/components/ui/animated-number"
-import { CompletionRing } from "@/components/dashboard/completion-ring"
+import { CompletionRing } from "@/components/ui/completion-ring"
 import { surfaceCardVariants } from "@/components/ui/surface-card"
 import { AlertTriangle, ExternalLink } from "lucide-react"
 import { cn } from "@/lib/utils"
