+++ components/ui/completion-ring.tsx
